pyproject.toml
[project]
name = "weaviate-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=4.0.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "weaviate>=0.1.2",
    "weaviate-client>=4.16.9",
]
